+++ libraries/core-react/src/components/Menu/Menu.tsx
@@ -5,7 +5,7 @@ import { useMenu } from './Menu.context'
 import { Paper } from '../Paper'
 import { MenuList } from './MenuList'
 import { useCombinedRefs } from '../../_common'
-import { useOutsideClick } from '../../_common/hooks'
+import { useOutsideClick } from '@hooks'
 import { menu as tokens } from './Menu.tokens'
 import type { State } from './Menu.context'
 import type { FocusTarget } from './Menu.types'
