+++ libraries/core-react/src/hooks/useOutsideClick.ts
